Fixes list styling across the website
Some of the list items in the markdown had an extra newline, causing it
to act like a paragraph. The design of the list items was done based on
that. I have removed all the extra newlines, and modified the design
such that it works with just list item text without a paragraph inside.